Publish DEPS for Chromium 12.0.742.131
Incrementing VERSION to 12.0.742.131

git-svn-id: svn://svn.chromium.org/chrome/branches/742/src@99363 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/video/omx_video_decode_engine.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include "base/callback.h"
-#include "base/scoped_ptr.h"
+#include "base/memory/scoped_ptr.h"
 #include "media/omx/omx_configurator.h"
 #include "media/video/video_decode_engine.h"
 #include "third_party/openmax/il/OMX_Component.h"
@@ -175,8 +175,8 @@ class OmxVideoDecodeEngine : public VideoDecodeEngine {
   scoped_refptr<VideoFrame> CreateOmxBufferVideoFrame(
       OMX_BUFFERHEADERTYPE* omx_buffer);
 
-  size_t width_;
-  size_t height_;
+  int width_;
+  int height_;
 
   MessageLoop* message_loop_;
